#IPEHR-43 EHR now changing status. Refactoring & tests

+++ src/pkg/docs/model/meta.go
@@ -6,5 +6,5 @@ type DocumentMeta struct {
 	TypeCode       types.DocumentType
 	StorageId      *[32]byte
 	DocIdEncrypted []byte
-	Timestamp      uint32
+	Timestamp      uint64
 }
